test(functional): move legacy fips tests to legacy folder
