--- tests/functional/endpoints/fips/test_cases_supported_variants.json
@@ -1,51 +1,0 @@
-[
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "ca-central-1-fips",
-    "signingRegion": "ca-central-1",
-    "hostname": "streams.dynamodb-fips.ca-central-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "us-east-1-fips",
-    "signingRegion": "us-east-1",
-    "hostname": "streams.dynamodb-fips.us-east-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "us-east-2-fips",
-    "signingRegion": "us-east-2",
-    "hostname": "streams.dynamodb-fips.us-east-2.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "us-west-1-fips",
-    "signingRegion": "us-west-1",
-    "hostname": "streams.dynamodb-fips.us-west-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "us-west-2-fips",
-    "signingRegion": "us-west-2",
-    "hostname": "streams.dynamodb-fips.us-west-2.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "us-gov-east-1-fips",
-    "signingRegion": "us-gov-east-1",
-    "hostname": "streams.dynamodb-fips.us-gov-east-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "streams.dynamodb",
-    "sdkId": "DynamoDB Streams",
-    "region": "us-gov-west-1-fips",
-    "signingRegion": "us-gov-west-1",
-    "hostname": "streams.dynamodb-fips.us-gov-west-1.amazonaws.com"
-  }
-]
