+++ tests/functional/endpoints/fips/test_cases_unsupported.json
@@ -1,86 +1,26 @@
 [
   {
-    "endpointPrefix": "dms",
-    "sdkId": "Database Migration Service",
-    "region": "dms-fips",
-    "signingRegion": "us-west-1",
-    "hostname": "dms-fips.us-west-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "dms",
-    "sdkId": "Database Migration Service",
-    "region": "dms-fips",
-    "signingRegion": "us-gov-west-1",
-    "hostname": "dms.us-gov-west-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "dms",
-    "sdkId": "Database Migration Service",
-    "region": "dms-fips",
-    "signingRegion": "us-iso-east-1",
-    "hostname": "dms.us-iso-east-1.c2s.ic.gov"
-  },
-  {
-    "endpointPrefix": "dms",
-    "sdkId": "Database Migration Service",
-    "region": "dms-fips",
-    "signingRegion": "us-isob-east-1",
-    "hostname": "dms.us-isob-east-1.sc2s.sgov.gov"
-  },
-  {
-    "endpointPrefix": "servicediscovery",
-    "sdkId": "ServiceDiscovery",
-    "region": "servicediscovery-fips",
-    "signingRegion": "ca-central-1",
-    "hostname": "servicediscovery-fips.ca-central-1.amazonaws.com"
+    "endpointPrefix": "iotsecuredtunneling",
+    "sdkId": "IoTSecureTunneling",
+    "region": "ap-east-1",
+    "hostname": "api.tunneling.iot-fips.ap-east-1.amazonaws.com"
   },
   {
     "endpointPrefix": "iotsecuredtunneling",
     "sdkId": "IoTSecureTunneling",
-    "region": "fips-us-gov-east-1",
-    "signingRegion": "us-gov-east-1",
-    "hostname": "api.tunneling.iot-fips.us-gov-east-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "fips-us-gov-west-1",
-    "signingRegion": "us-gov-west-1",
-    "hostname": "api.tunneling.iot-fips.us-gov-west-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "fips-ca-central-1",
-    "signingRegion": "ca-central-1",
+    "region": "ca-central-1",
     "hostname": "api.tunneling.iot-fips.ca-central-1.amazonaws.com"
   },
   {
     "endpointPrefix": "iotsecuredtunneling",
     "sdkId": "IoTSecureTunneling",
-    "region": "fips-us-east-1",
-    "signingRegion": "us-east-1",
-    "hostname": "api.tunneling.iot-fips.us-east-1.amazonaws.com"
+    "region": "cn-north-1",
+    "hostname": "iotsecuredtunneling-fips.cn-north-1.amazonaws.com.cn"
   },
   {
     "endpointPrefix": "iotsecuredtunneling",
     "sdkId": "IoTSecureTunneling",
-    "region": "fips-us-east-2",
-    "signingRegion": "us-east-2",
-    "hostname": "api.tunneling.iot-fips.us-east-2.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "fips-us-west-1",
-    "signingRegion": "us-west-1",
-    "hostname": "api.tunneling.iot-fips.us-west-1.amazonaws.com"
-  },
-  {
-    "endpointPrefix": "iotsecuredtunneling",
-    "sdkId": "IoTSecureTunneling",
-    "region": "fips-us-west-2",
-    "signingRegion": "us-west-2",
-    "hostname": "api.tunneling.iot-fips.us-west-2.amazonaws.com"
+    "region": "us-gov-east-1",
+    "hostname": "api.tunneling.iot-fips.us-gov-east-1.amazonaws.com"
   }
 ]
